Your solutions to 2018 codebreaker are publicly visible. Sorry for PR that's the only way I see to reach you...
Hey!

It's great you are posting a walkthrough to 2018 codebreaker challenge. But they have extended the deadline till January 6, and I know some people who are still working on this. You might want to keep the posts off the interwebs for a few days.

But otherwise great work.... I was going to write a walkthrough myself, but you did very superb job doing this.